Implemented two pass version
This version reduces memory requirements by making two passes over the input file. The first identifies all the ways (i.e. streets and pathways) that are candidates for the stress model. The second pass only stores the nodes that are referenced by these ways.  This takes twice as long to run but reduces the memory requirements by 90%.